pyproject.toml
[tool.poetry]
name = "discord_sound_streamer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.scripts]
start = "discord_sound_streamer.bot:start"

[tool.poetry.dependencies]
python = "~3.12"
aiohttp = "~3.10"
hikari = "2.1.0"
hikari-tanjun = "v2.17.6"
pydantic = "^2.9.2"
python-youtube = "^0.8.2"
httpx = "^0.23.0"
tenacity = "^8.0.1"
lavalink = { git = "https://github.com/zlalvani/Lavalink.py", rev = "zlalvani/fix-aiohttp-3-10" } # pinned because lavalink breaks with 3.10+
pydantic-settings = "^2.5.2"
attrs = "^24.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
mypy = "^0.931"
black = "^22.3.0"
isort = "^5.10.1"
pre-commit = "^2.18.1"
ipython = "^8.4.0"
jupyterlab = "^3.4.5"
pyright = "^1.1.383"
ruff = "^0.6.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict_optional = true
show_error_codes = true
show_column_numbers = true
warn_no_return = true
# disallow_any_unimported = true

# warn_unused_configs = true
# disallow_any_generics = true
# disallow_subclassing_any = true
# disallow_untyped_calls = true
# disallow_untyped_defs = true
# disallow_incomplete_defs = true
check_untyped_defs = true
# disallow_untyped_decorators = true
# no_implicit_optional = true
# warn_redundant_casts = true
# warn_return_any = true
strict_equality = true

[tool.pyright]
venvPath = "."
venv = ".venv"
typeCheckingMode = "standard"
reportMissingTypeStubs = "warning"

[tool.black]
line_length = 100

[tool.isort]
profile = "black"
